requirements.txt
scrapy
GitPython
schedule